ex18. Switch to CursorLoader.
Similar to network operations, data operations such as reading and writing to the pets table are expensive and time-consuming, so are database operations. Currently we are performing the ContentResolver query on the main thread in the CatalogActivity. But we don’t want to block the UI thread, so a loader allows us to do the operation on a background thread.